reorganized styles in common folder
