+++ src/radiology-tabs/common/listOrderDetails.scss
@@ -53,21 +53,3 @@
     margin-left: 1rem;
     margin-top: 2rem;
   }
-  .noOrdersDiv{
-    background-color: #fff;
-    text-align: center;
-    font-weight:bold;
-    width: 100%;
-  }
-  
-  .ordersTableToolbar{
-    position: static;
-    height: 3rem;
-    overflow: visible;
-    margin: 0;
-  }
-
-  .ordersToolbarSearchBar{
-    background-color:#f4f4f4 ;
-  }
-
