+++ src/radiology-tabs/common/radiologyProps.resource.ts
@@ -41,4 +41,5 @@ export interface ListOrdersDetailsProps {
 
 export interface OrderAction {
   actionName: string;
+  displayPosition: 0 | number;
 }
