+++ src/radiology-tabs/referred-test/referred-ordered.component.tsx
@@ -1,15 +1,10 @@
-import React, { useMemo, useState } from "react";
+import React from "react";
 import { useTranslation } from "react-i18next";
-import {
-  DataTableSkeleton,
-  OverflowMenu,
-  OverflowMenuItem,
-} from "@carbon/react";
+import { DataTableSkeleton } from "@carbon/react";
 import { useOrdersWorklist } from "../../hooks/useOrdersWorklist";
 import GroupedOrdersTable from "../common/groupedOrdersTable.component";
 
 export const ReferredTests: React.FC = () => {
-  const { t } = useTranslation();
   const { workListEntries, isLoading } = useOrdersWorklist("", "EXCEPTION");
 
   return isLoading ? (
