+++ src/radiology-tabs/review-ordered/review-ordered.component.tsx
@@ -16,7 +16,9 @@ export const Review: React.FC = () => {
         showStartButton={false}
         showActions={true}
         showOrderType={true}
-        actions={[{ actionName: "review-radilogy-report-dialog" }]}
+        actions={[
+          { actionName: "review-radilogy-report-dialog", displayPosition: 1 },
+        ]}
       />
     </div>
   );
