+++ src/radiology-tabs/test-ordered/radiology-instructions/radiology-instructions.component.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React from "react";
 import {
   Button,
   ModalBody,
